% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{accountUsage}
\alias{accountUsage}
\title{Show Account Usage}
\usage{
accountUsage(
  account = NULL,
  server = NULL,
  usageType = "hours",
  from = NULL,
  until = NULL,
  interval = NULL
)
}
\arguments{
\item{account, server}{Uniquely identify a remote server with either your
user \code{account}, the \code{server} name, or both. If neither are supplied, and
there are multiple options, you'll be prompted to pick one.

Use \code{\link[=accounts]{accounts()}} to see the full list of available options.}

\item{usageType}{Use metric to retreive (for example: "hours")}

\item{from}{Date range starting timestamp (Unix timestamp or relative time
delta such as "2d" or "3w").}

\item{until}{Date range ending timestamp (Unix timestamp or relative time
delta such as "2d" or "3w").}

\item{interval}{Summarization interval. Data points at intervals less then this
will be grouped. (Number of seconds or relative time delta e.g. "1h").}
}
\description{
Show account usage

Supported servers: ShinyApps servers
}
\note{
This function only works for ShinyApps servers.
}
