% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{addConnectServer}
\alias{addConnectServer}
\alias{addServer}
\alias{removeServer}
\alias{addServerCertificate}
\title{Server management}
\usage{
addConnectServer(url, name = NULL, certificate = NULL, quiet = FALSE)

addServer(url, name = NULL, certificate = NULL, quiet = FALSE)

removeServer(name = NULL)

addServerCertificate(name, certificate, quiet = FALSE)
}
\arguments{
\item{url}{Server's URL. Should look like \verb{http://servername/} or
\verb{http://servername:port/}.}

\item{name}{Optional nickname for the server. If none is given, the nickname
is inferred from the server's hostname.}

\item{certificate}{Optional. Either a path to certificate file or a
character vector containing the certificate's contents.}

\item{quiet}{Suppress output and prompts where possible.}
}
\description{
These functions manage the list of known servers:
\itemize{
\item \code{addServer()} registers a server.
\item \code{addConnectServer()} registers a Posit connect server. Once it has been
registered, you can connect to an account on the server using
\code{\link[=connectUser]{connectUser()}}.
\item \code{removeServer()} removes a server from the registry.
\item \code{addServerCertificate()} adds a certificate to a server.
}
}
\examples{
\dontrun{

# register a local server
addServer("http://myrsconnect/", "myserver")

# list servers
servers(local = TRUE)

# connect to an account on the server
connectUser(server = "myserver")
}
}
