% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{applicationConfigDir}
\alias{applicationConfigDir}
\title{Application Configuration Directory}
\usage{
applicationConfigDir()
}
\value{
A string containing the path of the configuration folder.
}
\description{
Returns the root path used to store per user configuration data. Does not
check old locations or create the path; use \code{rsconnectConfigDir} for
most cases.
}
\keyword{internal}
