% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundleFiles.R
\name{applyIgnorePatterns}
\alias{applyIgnorePatterns}
\title{Apply ignore patterns to a file list with hierarchical precedence}
\usage{
applyIgnorePatterns(file_list, patterns, current_dir)
}
\arguments{
\item{file_list}{List of file paths relative to current directory}

\item{patterns}{List of pattern objects (parent first, child last)}

\item{current_dir}{Current directory path}
}
\value{
Filtered file list with ignored files removed
}
\description{
Apply ignore patterns to a file list with hierarchical precedence
}
