% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundleFiles.R
\name{applyRscignorePatternsLegacy}
\alias{applyRscignorePatternsLegacy}
\title{Legacy .rscignore pattern application (deprecated)}
\usage{
applyRscignorePatternsLegacy(contents, dir)
}
\arguments{
\item{contents}{File contents to filter}

\item{dir}{Directory to check for .rscignore}
}
\value{
Filtered contents
}
\description{
Legacy .rscignore pattern application (deprecated)
}
