% Generated by roxygen2 (4.0.1): do not edit by hand
\name{authorizedUsers}
\alias{authorizedUsers}
\title{List authorized users for an application}
\usage{
authorizedUsers(appDir = getwd())
}
\arguments{
\item{appDir}{Directory containing application. Defaults to current working directory.}
}
\description{
List authorized users for an application
}

