% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{connectCloudUser}
\alias{connectCloudUser}
\title{Register account on Posit Connect Cloud}
\usage{
connectCloudUser(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started. Defaults to \code{TRUE} in
interactive sessions only. If a function is passed, it will be called
after the app is started, with the app URL as a parameter.}
}
\description{
\code{connectCloudUser()} connects your Posit Connect Cloud account to
the rsconnect package so that it can deploy and manage applications on
your behalf. It will open a browser window to authenticate, then prompt
you to create an account or select an account to use if you have multiple.

Supported servers: Posit Connect Cloud servers
}
\seealso{
Other Account functions: 
\code{\link{accounts}()},
\code{\link{connectApiUser}()},
\code{\link{setAccountInfo}()}
}
\concept{Account functions}
