% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{connectSPCSUser}
\alias{connectSPCSUser}
\title{Register account on Posit Connect in Snowpark Container Services}
\usage{
connectSPCSUser(
  account = NULL,
  server = NULL,
  snowflakeConnectionName,
  quiet = FALSE
)
}
\arguments{
\item{account}{A name for the account to connect.}

\item{server}{The server to connect to.}

\item{snowflakeConnectionName}{Name for the Snowflake connection parameters
stored in \code{connections.toml}.}

\item{quiet}{Whether or not to show messages and prompts while connecting the
account.}
}
\description{
\code{connectSPCSUser()} connects your Posit Connect account to the rsconnect
package so it can deploy and manage applications on your behalf.
Configure a
\href{https://docs.snowflake.com/en/developer-guide/snowflake-cli/connecting/configure-cli#location-of-the-toml-configuration-fil}{\code{connections.toml} file}
in the appropriate location.

Supported servers: Posit Connect servers
}
