% Generated by roxygen2 (4.0.1): do not edit by hand
\name{connectUser}
\alias{connectUser}
\title{Connect User Account}
\usage{
connectUser(server = NULL, username = "", quiet = FALSE)
}
\arguments{
\item{server}{The server to connect to. Optional if there is only one server
registered.}

\item{username}{An optional nickname for the account; applied only if the
account doesn't have a nickname already set on RStudio Connect.}

\item{quiet}{Whether or not to show messages and prompts while connecting
  the account.}
}
\description{
Connect an RStudio Connect user account to the package, so that it can be
used to deploy and manage applications on behalf of the account.
}
\details{
When this function is invoked, a web browser will be launched on
  RStudio Connect, where you will be prompted to enter your credentials. Upon
  successful authentication, your local installation of \pkg{rsconnect} and
  your RStudio Connect account will be paired, and you'll be able to deploy
  and manage applications using the package without further prompts for
  credentials.
}

