% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployAPI.R
\name{deployAPI}
\alias{deployAPI}
\title{Deploy a Plumber API}
\usage{
deployAPI(api, ...)
}
\arguments{
\item{api}{Path to the API project directory. Must contain either `entrypoint.R` or `plumber.R`}

\item{...}{Additional arguments to \code{\link{deployApp}}.}
}
\description{
Deploys an application consisting of plumber API routes. The given directory must contain a script returning a `plumb` object or a plumber API definition.
}
\details{
Deploy a plumber API definition by either supplying a directory containing `plumber.R` (an API definition)
    or `entrypoint.R` that returns a `plumb` object created by `plumber::plumb()`. See the plumber
    documentation for more information.
}
