% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploySite.R
\name{deploySite}
\alias{deploySite}
\title{Deploy a Website}
\usage{
deploySite(siteDir = getwd(), siteName = NULL, sourceCode = FALSE,
  render = FALSE, quiet = FALSE, lint = FALSE, ...)
}
\arguments{
\item{siteDir}{Directory containing website. Defaults to current
working directory.}

\item{siteName}{Name for the site (names must be unique within
an account). Defaults to the base name of the specified siteDir,
(or to a name provided by a custom site generation function).}

\item{sourceCode}{Should the site's source code be included in the
upload?}

\item{render}{\code{TRUE} to render the site before publishing.
This parameter is only applicable for \code{sourceCode = FALSE},
since rendering is done on the server when sourceCode is uploaded.}

\item{quiet}{Request that no status information be printed to the console
during the deployment.}

\item{lint}{Lint the project before initiating deployment, to identify
potentially problematic code?}

\item{...}{Additional arguments to pass to \code{\link{deployApp}}.}
}
\description{
Deploy an R Markdown website to a server.
}

