% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployTFModel.R
\name{deployTFModel}
\alias{deployTFModel}
\title{Deploy a Tensorflow saved model}
\usage{
deployTFModel(modelDir, ...)
}
\arguments{
\item{modelDir}{Path to the saved model directory. MUST contain `saved_model.pb`}

\item{...}{Additional arguments to \code{\link{deployApp}}.}
}
\description{
Deploys a directory containing a Tensorflow saved model file. A saved model directory might look like this:
`./1/`
`./1/saved_model.pb`
`./1/variables/`
`./1/variables/variables.data-00000-of-00001`
`./1/variables/variables.index`
}
\details{
Deploy a single Tensorflow saved model as a bundle. Should be passed a directory that contains the 
`saved_model.pb` file, as well as any variables and assets necessary to load the model.
}
