% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-deprec.R
\name{discoverServers}
\alias{discoverServers}
\title{Discover servers automatically}
\usage{
discoverServers(quiet = FALSE)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has never worked usefully, so has been removed.
}
\keyword{internal}
