% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lint}
\alias{lint}
\title{Lint a Project}
\usage{
lint(project, file = "")
}
\arguments{
\item{project}{Path to a project directory.}

\item{file}{Rmd file (can be empty)}
}
\description{
Takes the set of active linters (see \code{\link{addLinter}}), and applies
them to all files within a project.
}

