% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lint-utils.R
\name{makeLinterMessage}
\alias{makeLinterMessage}
\title{Construct a Linter Message}
\usage{
makeLinterMessage(header, content, lines)
}
\arguments{
\item{header}{A header message describing the linter.}

\item{content}{The content of the file that was linted.}

\item{lines}{The line numbers from \code{content} that contain lint.}
}
\description{
Pretty-prints a linter message. Primarily used as a helper
for constructing linter messages with \code{\link{linter}}.
}

