% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundleFiles.R
\name{matchDoubleAsteriskPattern}
\alias{matchDoubleAsteriskPattern}
\title{Match simple double-asterisk patterns}
\usage{
matchDoubleAsteriskPattern(file_path, pattern)
}
\arguments{
\item{file_path}{File path to match against}

\item{pattern}{Pattern object containing **}
}
\value{
TRUE if pattern matches, FALSE otherwise
}
\description{
Match simple double-asterisk patterns
}
