% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundleFiles.R
\name{matchGlobPattern}
\alias{matchGlobPattern}
\title{Match glob patterns}
\usage{
matchGlobPattern(file_path, pattern)
}
\arguments{
\item{file_path}{File path to match against}

\item{pattern}{Pattern object}
}
\value{
TRUE if pattern matches, FALSE otherwise
}
\description{
Match glob patterns
}
