% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{oldApplicationConfigDir}
\alias{oldApplicationConfigDir}
\title{Old Application Config Directory}
\usage{
oldApplicationConfigDir(appName)
}
\arguments{
\item{appName}{The application's name (connect or rsconnect)}
}
\value{
The old application configuration directory.
}
\description{
Returns the old application configuration directory used by rsconnect
0.8.24 and prior. These versions wrote configuration data to XDG compliant
locations, but CRAN policy has since further restricted the disk locations
that are permitted. See:
}
\details{
https://cran.r-project.org/web/packages/policies.html
}
\keyword{internal}
