% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumpstart.R
\name{open_jumpstart}
\alias{open_jumpstart}
\title{Open a jumpstart example from RStudio Connect}
\usage{
open_jumpstart(
  name,
  connect_url = Sys.getenv("CONNECT_URL", NA),
  destination = "~",
  open_project = TRUE
)
}
\arguments{
\item{name}{The name of the jumpstart example from an RStudio Connect
instance.}

\item{connect_url}{The URL of your Connect instance}

\item{destination}{Download path}

\item{open_project}{Boolean that controls whether to open the project after
download}
}
\description{
\code{pin_reactive_read()} and \code{pin_reactive_download()} wrap the results of
\code{\link[=pin_read]{pin_read()}} and \code{\link[=pin_download]{pin_download()}} into a Shiny reactive. This allows you to
use pinned data within your app, and have the results automatically
recompute when the pin is modified.
}
\examples{
if (FALSE) {
  open_jumpstart(
    name = "stock-report",
    connect_url = "https://connect.example.com",
    destination = "/home/user/projects",
    open_project = TRUE
  )
}
}
