% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundleFiles.R
\name{parseIgnoreFile}
\alias{parseIgnoreFile}
\title{Parse .rscignore file into pattern objects}
\usage{
parseIgnoreFile(directory_path)
}
\arguments{
\item{directory_path}{Path to directory containing .rscignore file}
}
\value{
List of pattern objects, or empty list if no .rscignore file
}
\description{
Parse .rscignore file into pattern objects
}
