% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundleFiles.R
\name{parseSinglePattern}
\alias{parseSinglePattern}
\title{Parse a single pattern line}
\usage{
parseSinglePattern(line)
}
\arguments{
\item{line}{Raw line from .rscignore file}
}
\value{
Pattern object or NULL if line should be skipped
}
\description{
Parse a single pattern line
}
