% Generated by roxygen2 (4.0.1): do not edit by hand
\name{registerUser}
\alias{registerUser}
\title{Register an existing user}
\usage{
registerUser(username, userId = 0, password = NULL, quiet = FALSE)
}
\arguments{
\item{username}{Username to register.}

\item{userId}{Numeric ID of user to register. Normally this is set to 0, in
which case the user ID is retrieved from the server.}

\item{password}{User's password. Can be \code{NULL}, in which case you will
be prompted to supply a password interactively. The password is not stored.}

\item{quiet}{Whether to print status messages.}
}
\description{
Registers an existing user account on an RStudio Connect server.
}
\details{
This function registers an existing user account on RStudio Connect
  with the \pkg{rsconnect} package. At least one account must be registered
  before applications can be deployed.
}

