% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{rsconnectConfigDir}
\alias{rsconnectConfigDir}
\title{rsconnect Configuration Directory}
\usage{
rsconnectConfigDir(subDir = NULL)
}
\arguments{
\item{subDir}{An optional subdirectory to be included as the last element of
the path.}
}
\value{
The path to the configuration directory.
}
\description{
Forms the path to a location on disk where user-level configuration data for
the package is stored.
}
\keyword{internal}
