% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{servers}
\alias{servers}
\alias{addConnectServer}
\alias{addServer}
\alias{removeServer}
\alias{serverInfo}
\alias{addServerCertificate}
\title{Server Management Functions}
\usage{
servers(local = FALSE)

addConnectServer(url, name = NULL, certificate = NULL, quiet = FALSE)

addServer(url, name = NULL, certificate = NULL, quiet = FALSE)

removeServer(name)

serverInfo(name)

addServerCertificate(name, certificate, quiet = FALSE)
}
\arguments{
\item{local}{Return only local servers (i.e. not \code{shinyapps.io})}

\item{url}{Server's URL. Should look like \verb{http://servername/} or
\verb{http://servername:port/}.}

\item{name}{Optional nickname for the server. If none is given, the nickname
is inferred from the server's hostname.}

\item{certificate}{Optional. Either a path to certificate file or a
character vector containing the certificate's contents.}

\item{quiet}{Suppress output and prompts where possible.}
}
\value{
\code{servers} returns a data frame with registered server names and URLs.
\code{serverInfo} returns a list with details for a particular server.
}
\description{
These functions manage the list of known servers:
\itemize{
\item \code{addServer()} registers a server. Once it has been registered, you
can connect to an account on the server using \code{\link[=connectUser]{connectUser()}}.
\item \code{servers()} lists known servers.
\item \code{serverInfo()} get information about a given server.
}

Servers for \code{shinyapps.io} and \code{posit.cloud} are always registered.
}
\examples{
\dontrun{

# register a local server
addServer("http://myrsconnect/", "myserver")

# list servers
servers(local = TRUE)

# connect to an account on the server
connectUser(server = "myserver")
}
}
