% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{setAccountInfo}
\alias{setAccountInfo}
\title{Register account on shinyapps.io or posit.cloud}
\usage{
setAccountInfo(name, token, secret, server = "shinyapps.io")
}
\arguments{
\item{name}{Name of account to save or remove}

\item{token}{User token for the account}

\item{secret}{User secret for the account}

\item{server}{Server to associate account with.}
}
\description{
Configure a ShinyApps or Posit Cloud account for publishing from this system.
}
\examples{
\dontrun{

# register an account
setAccountInfo("user", "token", "secret")

# remove the same account
removeAccount("user")
}

}
\seealso{
Other Account functions: 
\code{\link{accounts}()},
\code{\link{connectApiUser}()}
}
\concept{Account functions}
