% Generated by roxygen2 (4.0.1): do not edit by hand
\name{showLogs}
\alias{showLogs}
\title{Show Application Logs}
\usage{
showLogs(appDir = getwd(), appName = NULL, account = NULL, entries = 50)
}
\arguments{
\item{appDir}{The directory where the local copy of the application
resides.}

\item{appName}{The name of the application to show logs for. May be omitted
if only one application deployment was made from \code{appDir}.}

\item{account}{The account under which the application was deployed. May be
omitted if only one account is registered on the system.}

\item{entries}{The number of log entries to show. Defaults to 50 entries.}
}
\description{
Show the logs for a deployed application.
}

