% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{showMetrics}
\alias{showMetrics}
\title{Show Application Metrics}
\usage{
showMetrics(
  metricSeries,
  metricNames,
  appDir = getwd(),
  appName = NULL,
  account = NULL,
  server = "shinyapps.io",
  from = NULL,
  until = NULL,
  interval = NULL
)
}
\arguments{
\item{metricSeries}{Metric series to query. Refer to the
\href{https://docs.posit.co/shinyapps.io/metrics.html#ApplicationMetrics}{shinyapps.io documentation}
for available series.}

\item{metricNames}{Metric names in the series to query. Refer to the
\href{https://docs.posit.co/shinyapps.io/metrics.html#ApplicationMetrics}{shinyapps.io documentation}
for available metrics.}

\item{appDir}{Path to an application directory. Defaults to the
working directory.}

\item{appName}{Name of application (names must be unique within an account).
If not supplied on the first deploy, it will be generated from the base
name of \code{appDir} and \code{appTitle}, if supplied. On subsequent deploys,
it will use the previously stored value.}

\item{account, server}{Uniquely identify a remote server with either your
user \code{account}, the \code{server} name, or both. Use \code{\link[=accounts]{accounts()}} to see the
full list of available options.}

\item{from}{Date range starting timestamp (Unix timestamp or relative time
delta such as "2d" or "3w").}

\item{until}{Date range ending timestamp (Unix timestamp or relative time
delta such as "2d" or "3w").}

\item{interval}{Summarization interval. Data points at intervals less then this
will be grouped. (Relative time delta e.g. "120s" or "1h" or "30d").}
}
\description{
Show application metrics of a currently deployed application.
This function only works for ShinyApps servers.
}
