% Generated by roxygen2 (4.0.1): do not edit by hand
\name{showProperties}
\alias{showProperties}
\title{Show Application property}
\usage{
showProperties(appDir = getwd(), appName = NULL, account = NULL)
}
\arguments{
\item{appName}{Name of application}

\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{account}{Account name. If a single account is registered on the
system then this parameter can be omitted.}
}
\description{
Show application propreties of a currently deployed application
}

