% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{showUsage}
\alias{showUsage}
\title{Show Application Usage}
\usage{
showUsage(
  appDir = getwd(),
  appName = NULL,
  account = NULL,
  server = NULL,
  usageType = "hours",
  from = NULL,
  until = NULL,
  interval = NULL
)
}
\arguments{
\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application}

\item{account, server}{Uniquely identify a remote server with either your
user \code{account}, the \code{server} name, or both. If neither are supplied, and
there are multiple options, you'll be prompted to pick one.

Use \code{\link[=accounts]{accounts()}} to see the full list of available options.}

\item{usageType}{Use metric to retreive (for example: "hours")}

\item{from}{Date range starting timestamp (Unix timestamp or relative time
delta such as "2d" or "3w").}

\item{until}{Date range ending timestamp (Unix timestamp or relative time
delta such as "2d" or "3w").}

\item{interval}{Summarization interval. Data points at intervals less then this
will be grouped. (Relative time delta e.g. "120s" or "1h" or "30d").}
}
\description{
Show application usage of a currently deployed application

Supported servers: ShinyApps servers
}
\note{
This function only works for ShinyApps servers.
}
