% Generated by roxygen2 (4.0.2): do not edit by hand
\name{accountUsage}
\alias{accountUsage}
\title{Show Account Usage}
\usage{
accountUsage(account = NULL, usageType = "hours", from = NULL,
  until = NULL, interval = NULL)
}
\arguments{
\item{account}{Account name. If a single account is registered on the
system then this parameter can be omitted.}

\item{usageType}{}

\item{from}{}

\item{until}{}

\item{interval}{}
}
\description{
Show account usage
}

