% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/usage.R
\name{accountUsage}
\alias{accountUsage}
\title{Show Account Usage}
\usage{
accountUsage(account = NULL, server = NULL, usageType = "hours",
  from = NULL, until = NULL, interval = NULL)
}
\arguments{
\item{account}{Account name. If a single account is registered on the
system then this parameter can be omitted.}

\item{server}{Server name. Required only if you use the same account name on
multiple servers.}

\item{usageType}{Use metric to retreive (for example: "hours")}

\item{from}{Date range starting timestamp (Unix timestamp or relative time
delta such as "2d" or "3w").}

\item{until}{Date range ending timestamp (Unix timestamp or relative time
delta such as "2d" or "3w").}

\item{interval}{Summarization interval. Data points at intervals less then this
will be grouped. (Number of seconds or relative time delta e.g. "1h").}
}
\description{
Show account usage
}
\note{
This function only works for ShinyApps servers.
}

