% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{addAuthorizedUser}
\alias{addAuthorizedUser}
\title{Add authorized user to application}
\usage{
addAuthorizedUser(
  email,
  appDir = getwd(),
  appName = NULL,
  contentId = NULL,
  account = NULL,
  server = NULL,
  sendEmail = NULL,
  emailMessage = NULL
)
}
\arguments{
\item{email}{Email address of user to add.}

\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application.}

\item{contentId}{On Posit Connect Cloud, the content ID to manage, taken from
the content URL
(\code{https://connect.posit.cloud/{account}/content/{contentId}}). When
supplied, \code{appDir} and \code{appName} are ignored and no local
deployment record is required. Not supported on shinyapps.io.}

\item{account, server}{Uniquely identify a remote server with either your
user \code{account}, the \code{server} name, or both. If neither are supplied, and
there are multiple options, you'll be prompted to pick one.

Use \code{\link[=accounts]{accounts()}} to see the full list of available options.}

\item{sendEmail}{Send an email letting the user know the application
has been shared with them.}

\item{emailMessage}{Optional character vector of length 1 containing a
custom message to send in email invitation. Defaults to NULL, which
will use default invitation message.}
}
\description{
Add authorized user to application

Supported servers: ShinyApps, Posit Connect Cloud
}
\note{
This function works for ShinyApps and Posit Connect Cloud. On Posit
Connect Cloud, the content's account must be an organization account.
The \code{sendEmail} argument is ignored on Posit Connect Cloud;
PCC always sends an invitation email.

On Posit Connect Cloud, the content is resolved from the local deployment
record under \code{appDir}, which defaults to the working directory. Pass
\code{appDir} to point at the project directory that contains the
\code{rsconnect/} deployment record. \code{appName} selects among multiple records in
the same directory. Alternatively, pass \code{contentId} to target the
content directly, without a local deployment record.
}
\seealso{
\code{\link[=removeAuthorizedUser]{removeAuthorizedUser()}} and \code{\link[=showUsers]{showUsers()}}
}
