% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addAuthorizedUser}
\alias{addAuthorizedUser}
\title{Add authorized user for application}
\usage{
addAuthorizedUser(username, password = NULL, appDir = getwd())
}
\arguments{
\item{username}{The user name.}

\item{password}{The password.}

\item{appDir}{Directory containing application. Defaults to current working directory.}
}
\description{
Add authorized user for application
}
\examples{
\dontrun{

# add a user (prompts for password)
addAuthorizedUser("andy")

# add a user using supplied password
addAuthorizedUser("tareef", "MrShiny45")

}
}
\seealso{
\code{\link{removeAuthorizedUser}} and \code{\link{authorizedUsers}}
}

