% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auth.R
\name{addAuthorizedUser}
\alias{addAuthorizedUser}
\title{Add authorized user for application}
\usage{
addAuthorizedUser(username, password = NULL, appPath = getwd())
}
\arguments{
\item{username}{The user name.}

\item{password}{The password.}

\item{appPath}{Directory or file that was deployed. Defaults to current
working directory.}
}
\description{
Add authorized user for application
}
\examples{
\dontrun{

# add a user (prompts for password)
addAuthorizedUser("andy")

# add a user using supplied password
addAuthorizedUser("tareef", "MrShiny45")

}
}
\seealso{
\code{\link{removeAuthorizedUser}} and \code{\link{authorizedUsers}}
}

