% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auth.R
\name{addAuthorizedUser}
\alias{addAuthorizedUser}
\title{Add authorized user to application}
\usage{
addAuthorizedUser(email, appDir = getwd(), appName = NULL, account = NULL,
  sendEmail = TRUE)
}
\arguments{
\item{email}{Email address of user to add.}

\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application.}

\item{account}{Account name. If a single account is registered on the
system then this parameter can be omitted.}
}
\description{
Add authorized user to application
}
\seealso{
\code{\link{removeAuthorizedUser}} and \code{\link{showUsers}}
}

