% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers-deprec.R
\name{addConnectServer}
\alias{addConnectServer}
\title{Add a server}
\usage{
addConnectServer(url, name = NULL, certificate = NULL, quiet = FALSE)
}
\arguments{
\item{url}{URL for the server. Can be a bare hostname like
\code{connect.mycompany.com} or a url like \verb{http://posit.mycompany.com/connect}.}

\item{name}{Server name. If omitted, the server hostname is used.}

\item{certificate}{Optional. Either a path to certificate file or a
character vector containing the certificate's contents.}

\item{quiet}{Suppress output and prompts where possible.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function does the same thing as \code{\link[=addServer]{addServer()}} so has been removed.
}
\keyword{internal}
