% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auth.R
\name{authorizedUsers}
\alias{authorizedUsers}
\title{List authorized users for an application}
\usage{
authorizedUsers(appDir = getwd())
}
\arguments{
\item{appDir}{Directory containing application. Defaults to current working directory.}
}
\description{
List authorized users for an application
}

