% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{authorizedUsers}
\alias{authorizedUsers}
\title{(Deprecated) List authorized users for an application}
\usage{
authorizedUsers(appDir = getwd())
}
\arguments{
\item{appDir}{Directory containing application. Defaults to current working
directory.}
}
\description{
(Deprecated) List authorized users for an application
}

