% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{connectCloudClientCredentials}
\alias{connectCloudClientCredentials}
\title{Register a Posit Connect Cloud account using OAuth client credentials}
\usage{
connectCloudClientCredentials(
  clientId,
  clientSecret,
  accountName,
  name = NULL,
  quiet = FALSE
)
}
\arguments{
\item{clientId}{The OAuth client ID issued for a Posit Connect Cloud
service account.}

\item{clientSecret}{The OAuth client secret paired with \code{clientId}.}

\item{accountName}{The Posit Connect Cloud account name to publish to. The
credentials must grant publish permission on this account.}

\item{name}{The local name to record the account under. Defaults to
\code{accountName}.}

\item{quiet}{Whether or not to show messages while connecting the account.}
}
\description{
\code{connectCloudClientCredentials()} registers a Posit Connect Cloud account
using an OAuth 2.0 \code{client_credentials} grant provided by the Posit Cloud
auth service (https://login.posit.cloud/identity/credentials). Use this
function to authenticate in non-interactive contexts.

Supported servers: Posit Connect Cloud servers
}
\seealso{
Other Account functions:
\code{\link[=accounts]{accounts()}},
\code{\link[=connectApiUser]{connectApiUser()}},
\code{\link[=connectCloudUser]{connectCloudUser()}},
\code{\link[=setAccountInfo]{setAccountInfo()}}
}
\concept{Account functions}
