% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{hr}
\alias{hr}
\title{Generate a line with embedded message}
\usage{
hr(message = "", n = 80)
}
\arguments{
\item{message}{A string (single-line message).}

\item{n}{The total number length of the generated string --
the message is padded with \code{#} up to length \code{n}.}
}
\description{
Generates a message, surrounded with \code{#}, that extends
up to length \code{n}.
}

