% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lint-framework.R
\name{lint}
\alias{lint}
\title{Lint a Project}
\usage{
lint(project, files = NULL)
}
\arguments{
\item{project}{Path to a project directory.}

\item{files}{Specific files to lint. Can be NULL, in which case all
the files in the directory will be linted.}
}
\description{
Takes the set of active linters (see \code{\link{addLinter}}), and applies
them to all files within a project.
}

