% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundleFiles.R
\name{listBundleFiles}
\alias{listBundleFiles}
\title{List Files to be Bundled}
\usage{
listBundleFiles(appDir)
}
\arguments{
\item{appDir}{Directory containing the application.}
}
\value{
Returns a list containing the following elements:
\itemize{
\item \code{totalFiles}: Total number of files.
\item \code{totalSize}: Total size of the files (in bytes).
\item \code{contents}: Paths to bundle, relative to \code{appDir}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

\code{listBundleFiles()} has been superseded in favour of \code{\link[=listDeploymentFiles]{listDeploymentFiles()}}.

Given a directory containing an application, returns the names of the files
that by default will be bundled in the application. It works similarly to
a recursive directory listing from \code{\link[=list.files]{list.files()}} but enforces bundle sizes
as described in \code{\link[=listDeploymentFiles]{listDeploymentFiles()}}
}
\keyword{internal}
