% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auth.R
\name{removeAuthorizedUser}
\alias{removeAuthorizedUser}
\title{Remove authorized user from an application}
\usage{
removeAuthorizedUser(username, appPath = getwd())
}
\arguments{
\item{username}{The user name.}

\item{appPath}{Directory or file that was deployed. Defaults to current
working directory.}
}
\description{
Remove authorized user from an application
}
\examples{
\dontrun{

# remove user
removeAuthorizedUser("andy")

}
}
\seealso{
\code{\link{addAuthorizedUser}} and \code{\link{authorizedUsers}}
}

