% Generated by roxygen2 (4.0.2): do not edit by hand
\name{removeAuthorizedUser}
\alias{removeAuthorizedUser}
\title{Remove authorized user from an application}
\usage{
removeAuthorizedUser(username, appDir = getwd())
}
\arguments{
\item{username}{The user name.}

\item{appDir}{Directory containing application. Defaults to current working directory.}
}
\description{
Remove authorized user from an application
}
\examples{
\dontrun{

# remove user
removeAuthorizedUser("andy")

}
}
\seealso{
\code{\link{addAuthorizedUser}} and \code{\link{authorizedUsers}}
}

