% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeAuthorizedUser}
\alias{removeAuthorizedUser}
\title{Remove authroized user from an application}
\usage{
removeAuthorizedUser(username, appDir = getwd())
}
\arguments{
\item{username}{}

\item{appDir}{Directory containing application. Defaults to current working directory.}
}
\description{
Remove authroized user from an application
}
\examples{
\dontrun{

# remove user
removeAuthroizedUser("andy")

}
}
\seealso{
\code{\link{addAuthorizedUser}} and \code{\link{authorizedUsers}}
}

