% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{removeAuthorizedUser}
\alias{removeAuthorizedUser}
\title{Remove authorized user from an application}
\usage{
removeAuthorizedUser(
  user,
  appDir = getwd(),
  appName = NULL,
  account = NULL,
  server = NULL
)
}
\arguments{
\item{user}{The user to remove. Can be id or email address.}

\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application.}

\item{account, server}{Uniquely identify a remote server with either your
user \code{account}, the \code{server} name, or both. If neither are supplied, and
there are multiple options, you'll be prompted to pick one.

Use \code{\link[=accounts]{accounts()}} to see the full list of available options.}
}
\description{
Remove authorized user from an application
}
\note{
This function works only for ShinyApps servers.
}
\seealso{
\code{\link[=addAuthorizedUser]{addAuthorizedUser()}} and \code{\link[=showUsers]{showUsers()}}
}
