% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{resendInvitation}
\alias{resendInvitation}
\title{Resend invitation for invited users of an application}
\usage{
resendInvitation(
  invite,
  regenerate = FALSE,
  appDir = getwd(),
  appName = NULL,
  contentId = NULL,
  account = NULL,
  server = NULL
)
}
\arguments{
\item{invite}{The invitation to resend. Can be id or email address.}

\item{regenerate}{Regenerate the invite code. Can be helpful is the
invitation has expired.}

\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application.}

\item{contentId}{On Posit Connect Cloud, the content ID to manage, taken from
the content URL
(\code{https://connect.posit.cloud/{account}/content/{contentId}}). When
supplied, \code{appDir} and \code{appName} are ignored and no local
deployment record is required. Not supported on shinyapps.io.}

\item{account, server}{Uniquely identify a remote server with either your
user \code{account}, the \code{server} name, or both. If neither are supplied, and
there are multiple options, you'll be prompted to pick one.

Use \code{\link[=accounts]{accounts()}} to see the full list of available options.}
}
\description{
Resend invitation for invited users of an application

Supported servers: ShinyApps, Posit Connect Cloud
}
\note{
This function works for ShinyApps and Posit Connect Cloud. The
invitation can be selected by id or email address. On Posit Connect Cloud,
the \code{regenerate} argument has no effect.

On Posit Connect Cloud, the content is resolved from the local deployment
record under \code{appDir}, which defaults to the working directory. Pass
\code{appDir} to point at the project directory that contains the
\code{rsconnect/} deployment record. \code{appName} selects among multiple records in
the same directory. Alternatively, pass \code{contentId} to target the
content directly, without a local deployment record.
}
\seealso{
\code{\link[=showInvited]{showInvited()}}
}
