% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{resendInvitation}
\alias{resendInvitation}
\title{Resend invitation for invited users of an application}
\usage{
resendInvitation(
  invite,
  regenerate = FALSE,
  appDir = getwd(),
  appName = NULL,
  account = NULL,
  server = NULL
)
}
\arguments{
\item{invite}{The invitation to resend. Can be id or email address.}

\item{regenerate}{Regenerate the invite code. Can be helpful is the
invitation has expired.}

\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application.}

\item{account}{Account name. If a single account is registered on the
system then this parameter can be omitted.}

\item{server}{Server name. Required only if you use the same account name on
multiple servers.}
}
\description{
Resend invitation for invited users of an application
}
\note{
This function works only for ShinyApps servers.
}
\seealso{
\code{\link[=showInvited]{showInvited()}}
}
