% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsconnect-package.R
\docType{package}
\name{rsconnect-package}
\alias{rsconnect}
\alias{rsconnect-package}
\title{rsconnect: Deploy Docs, Apps, and APIs to 'Posit Connect', 'shinyapps.io', and 'RPubs'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Programmatic deployment interface for 'RPubs', 'shinyapps.io', and 'Posit Connect'. Supported content types include R Markdown documents, Shiny applications, Plumber APIs, plots, and static web content.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/rsconnect}
  \item Report bugs at \url{https://github.com/rstudio/rsconnect/issues}
}

}
\author{
\strong{Maintainer}: Aron Atkins \email{aron@posit.co}

Authors:
\itemize{
  \item Toph Allen
  \item Hadley Wickham
  \item Jonathan McPherson
  \item JJ Allaire
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
