% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servers.R
\name{servers}
\alias{servers}
\alias{serverInfo}
\title{Server metadata}
\usage{
servers(local = FALSE)

serverInfo(name = NULL)
}
\arguments{
\item{local}{Return only local servers? (i.e. not automatically registered
cloud servers)}

\item{name}{Server name. If omitted, you'll be prompted to pick a server.}
}
\value{
\code{servers()} returns a data frame with registered server names and URLs.
\code{serverInfo()} returns a list with details for a particular server.
}
\description{
\code{servers()} lists all known servers; \code{serverInfo()} gets metadata about
a specific server. Cloud server \code{shinyapps.io} is always automatically
registered and available.

Supported servers: All servers
}
\examples{
# List all registered servers
servers()

# Get information about a server
serverInfo("shinyapps.io")
}
