% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accounts.R
\name{setAccountInfo}
\alias{setAccountInfo}
\title{Set ShinyApps Account Info}
\usage{
setAccountInfo(name, token, secret)
}
\arguments{
\item{name}{Name of account to save or remove}

\item{token}{User token for the account}

\item{secret}{User secret for the account}
}
\description{
Configure a ShinyApps account for publishing from this system.
}
\examples{
\dontrun{

# register an account
setAccountInfo("user", "token", "secret")

# remove the same account
removeAccount("user")
}
}

