% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setProperty}
\alias{setProperty}
\title{Set Application property}
\usage{
setProperty(propertyName, propertyValue, appDir = getwd(), appName = NULL,
  account = NULL, force = FALSE)
}
\arguments{
\item{propertyName}{Name of property to set}

\item{propertyValue}{Nalue to set property to}

\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application}

\item{account}{Account name. If a single account is registered on the
system then this parameter can be omitted.}

\item{force}{Forcibly set the property}
}
\description{
Set a property on currently deployed application
}
\examples{
\dontrun{

# set instance size for an application
setProperty("application.instances.count", 1)

# disable application package cache
setProperty("application.package.cache", FALSE)

}
}

