\name{shinyapps-package}
\alias{shinyapps-package}
\alias{shinyapps}
\docType{package}
\title{
Interface for publishing applications to shinyapps.io
}
\description{
Interface for publishing applications to shinyapps.io
}
\details{
\tabular{ll}{
Package: \tab shinyapps\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-09-27\cr
License: \tab GPL-2\cr
}



}
\author{
JJ Allaire

Maintainer: JJ Allaire <jj@rstudio.com>
}
\references{

}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[shiny:shiny-package]{shiny}} ~~
}
\examples{

}
