\name{shinyapps-package}
\alias{shinyapps-package}
\alias{shinyapps}
\docType{package}
\title{
Interface to ShinyApps
}
\description{
Interface for deploying \pkg{shiny} applications to the ShinyApps service.
}
\details{
\tabular{ll}{
Package: \tab shinyapps\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2013-10-03\cr
License: \tab GPL-2\cr
}

}
\author{
JJ Allaire

Maintainer: JJ Allaire <jj@rstudio.com>
}

\keyword{ package }
\seealso{
\link[shiny:shiny-package]{shiny}
}

