% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{showInvited}
\alias{showInvited}
\title{List invited users for an application}
\usage{
showInvited(
  appDir = getwd(),
  appName = NULL,
  contentId = NULL,
  account = NULL,
  server = NULL
)
}
\arguments{
\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application.}

\item{contentId}{On Posit Connect Cloud, the content ID to manage, taken from
the content URL
(\code{https://connect.posit.cloud/{account}/content/{contentId}}). When
supplied, \code{appDir} and \code{appName} are ignored and no local
deployment record is required. Not supported on shinyapps.io.}

\item{account, server}{Uniquely identify a remote server with either your
user \code{account}, the \code{server} name, or both. If neither are supplied, and
there are multiple options, you'll be prompted to pick one.

Use \code{\link[=accounts]{accounts()}} to see the full list of available options.}
}
\description{
List invited users for an application

Supported servers: ShinyApps, Posit Connect Cloud
}
\note{
This function works for ShinyApps and Posit Connect Cloud. On Posit
Connect Cloud, the \code{link} column is always \code{NA} because the
accept link is only emailed to the recipient and is never returned by
the API.

On Posit Connect Cloud, the content is resolved from the local deployment
record under \code{appDir}, which defaults to the working directory. Pass
\code{appDir} to point at the project directory that contains the
\code{rsconnect/} deployment record. \code{appName} selects among multiple records in
the same directory. Alternatively, pass \code{contentId} to target the
content directly, without a local deployment record.
}
\seealso{
\code{\link[=addAuthorizedUser]{addAuthorizedUser()}} and \code{\link[=showUsers]{showUsers()}}
}
