% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{showInvited}
\alias{showInvited}
\title{List invited users for an application}
\usage{
showInvited(appDir = getwd(), appName = NULL, account = NULL, server = NULL)
}
\arguments{
\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application.}

\item{account, server}{Uniquely identify a remote server with either your
user \code{account}, the \code{server} name, or both. If neither are supplied, and
there are multiple options, you'll be prompted to pick one.

Use \code{\link[=accounts]{accounts()}} to see the full list of available options.}
}
\description{
List invited users for an application
}
\note{
This function works only for ShinyApps servers.
}
\seealso{
\code{\link[=addAuthorizedUser]{addAuthorizedUser()}} and \code{\link[=showUsers]{showUsers()}}
}
