% Generated by roxygen2 (4.0.2): do not edit by hand
\name{showLogs}
\alias{showLogs}
\title{Show Application Logs}
\usage{
showLogs(appDir = getwd(), appName = NULL, account = NULL, entries = 50,
  streaming = FALSE)
}
\arguments{
\item{appDir}{The directory where the local copy of the application
resides.}

\item{appName}{The name of the application to show logs for. May be omitted
if only one application deployment was made from \code{appDir}.}

\item{account}{The account under which the application was deployed. May be
omitted if only one account is registered on the system.}

\item{entries}{The number of log entries to show. Defaults to 50 entries.}

\item{streaming}{Whether to stream the logs. If \code{TRUE}, then the
function does not return; instead, log entries are written to the console
as they are made, until R is interrupted. Defaults to \code{FALSE}.}
}
\description{
Show the logs for a deployed application.
}

