% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{showLogs}
\alias{showLogs}
\title{Show Application Logs}
\usage{
showLogs(
  appPath = getwd(),
  appFile = NULL,
  appName = NULL,
  account = NULL,
  server = NULL,
  entries = 50,
  streaming = FALSE
)
}
\arguments{
\item{appPath}{The path to the directory or file that was deployed.}

\item{appFile}{The path to the R source file that contains the application
(for single file applications).}

\item{appName}{The name of the application to show logs for. May be omitted
if only one application deployment was made from \code{appPath}.}

\item{account}{The account under which the application was deployed. May be
omitted if only one account is registered on the system.}

\item{server}{Server name. Required only if you use the same account name on
multiple servers.}

\item{entries}{The number of log entries to show. Defaults to 50 entries.}

\item{streaming}{Whether to stream the logs. If \code{TRUE}, then the
function does not return; instead, log entries are written to the console
as they are made, until R is interrupted. Defaults to \code{FALSE}.}
}
\description{
Show the logs for a deployed ShinyApps application.
}
\note{
This function only uses the \code{libcurl} transport, and works only for
ShinyApps servers.
}
