% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureApp.R
\name{showProperties}
\alias{showProperties}
\title{Show Application property}
\usage{
showProperties(
  appPath = getwd(),
  appName = NULL,
  account = NULL,
  server = NULL
)
}
\arguments{
\item{appPath}{Directory or file that was deployed. Defaults to current
working directory.}

\item{appName}{Name of application}

\item{account, server}{Uniquely identify a remote server with either your
user \code{account}, the \code{server} name, or both. If neither are supplied, and
there are multiple options, you'll be prompted to pick one.

Use \code{\link[=accounts]{accounts()}} to see the full list of available options.}
}
\description{
Show properties of an application deployed to ShinyApps.
}
\note{
This function works only for ShinyApps servers.
}
