% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/configureApp.R
\name{showProperties}
\alias{showProperties}
\title{Show Application property}
\usage{
showProperties(appPath = getwd(), appName = NULL, account = NULL)
}
\arguments{
\item{appPath}{Directory or file that was deployed. Defaults to current
working directory.}

\item{appName}{Name of application}

\item{account}{Account name. If a single account is registered on the
  system then this parameter can be omitted.}
}
\description{
Show propreties of an application deployed to ShinyApps.
}
\note{
This function works only for ShinyApps servers.
}

