% Generated by roxygen2 (4.0.2): do not edit by hand
\name{showUsage}
\alias{showUsage}
\title{Show Application Usage}
\usage{
showUsage(appDir = getwd(), appName = NULL, account = NULL,
  usageType = "hours", from = NULL, until = NULL, interval = NULL)
}
\arguments{
\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application}

\item{account}{Account name. If a single account is registered on the
system then this parameter can be omitted.}

\item{usageType}{}

\item{from}{}

\item{until}{}

\item{interval}{}
}
\description{
Show application usage of a currently deployed application
}

