% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{showUsers}
\alias{showUsers}
\title{List authorized users for an application}
\usage{
showUsers(appDir = getwd(), appName = NULL, account = NULL, server = NULL)
}
\arguments{
\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application.}

\item{account, server}{Uniquely identify a remote server with either your
user \code{account}, the \code{server} name, or both. If neither are supplied, and
there are multiple options, you'll be prompted to pick one.

Use \code{\link[=accounts]{accounts()}} to see the full list of available options.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

List authorized users for an application

Supported servers: ShinyApps, Posit Connect Cloud
}
\note{
This function works for ShinyApps and Posit Connect Cloud.

On Posit Connect Cloud, the content is resolved from the local deployment
record in \code{appDir}; the call must run from the directory that contains
the \code{rsconnect/} deployment record. \code{appName} selects among
multiple records in the same directory.
}
\seealso{
\code{\link[=addAuthorizedUser]{addAuthorizedUser()}} and \code{\link[=showInvited]{showInvited()}}
}
