% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.R
\name{showUsers}
\alias{showUsers}
\title{List authorized users for an application}
\usage{
showUsers(appDir = getwd(), appName = NULL, account = NULL,
  server = NULL)
}
\arguments{
\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application.}

\item{account}{Account name. If a single account is registered on the
system then this parameter can be omitted.}

\item{server}{Server name. Required only if you use the same account name on
multiple servers.}
}
\description{
List authorized users for an application
}
\note{
This function works only for ShinyApps servers.
}
\seealso{
\code{\link{addAuthorizedUser}} and \code{\link{showInvited}}
}

