% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{syncAppMetadata}
\alias{syncAppMetadata}
\title{Sync Application Metadata}
\usage{
syncAppMetadata(appPath)
}
\arguments{
\item{appPath}{The path to the directory or file that was deployed.}
}
\description{
Update the metadata for requested application across all deployments
}
\note{
This function does not update metadata for Shiny and rpubs apps
}
