% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{syncAppMetadata}
\alias{syncAppMetadata}
\title{Update deployment records}
\usage{
syncAppMetadata(appPath = ".")
}
\arguments{
\item{appPath}{The path to the directory or file that was deployed.}
}
\description{
Update the deployment records for applications published to Posit Connect.
This updates application title and URL, and deletes records for deployments
where the application has been deleted on the server.
}
