% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tasks.R
\name{taskLog}
\alias{taskLog}
\title{Show task log}
\usage{
taskLog(taskId, account = NULL, server = NULL, output = NULL)
}
\arguments{
\item{taskId}{Task Id}

\item{account}{Account name. If a single account is registered on the system
then this parameter can be omitted.}

\item{server}{Server name. Required only if you use the same account name on
multiple servers (see \code{\link{servers}})}

\item{output}{Where to write output. Valid values are \code{NULL} or \code{stderr}}
}
\description{
Writes the task log for the given task
}
\examples{
\dontrun{

# write task log to stdout
taskLog(12345)

# write task log to stderr
taskLog(12345, output="stderr")

}
}
\seealso{
\code{\link{tasks}}
}

