% Generated by roxygen2 (4.0.2): do not edit by hand
\name{terminateApp}
\alias{terminateApp}
\title{Terminate an Application}
\usage{
terminateApp(appName, account = NULL, quiet = FALSE)
}
\arguments{
\item{appName}{Name of application to terminate}

\item{account}{Account name. If a single account is registered on the
system then this parameter can be omitted.}

\item{quiet}{Request that no status information be printed to the console
during the termination.}
}
\description{
Terminate an application currently running on ShinyApps.
}
\examples{
\dontrun{

# terminate an application
terminateApp("myapp")
}
}
\seealso{
\code{\link{applications}}, \code{\link{deployApp}}, and
  \code{\link{restartApp}}
}

