% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unsetProperty}
\alias{unsetProperty}
\title{Unset Application property}
\usage{
unsetProperty(propertyName, appDir = getwd(), appName = NULL,
  account = NULL, force = FALSE)
}
\arguments{
\item{propertyName}{Name of property to unset}

\item{appDir}{Directory containing application. Defaults to
current working directory.}

\item{appName}{Name of application}

\item{account}{Account name. If a single account is registered on the
system then this parameter can be omitted.}

\item{force}{Forcibly unset the property}
}
\description{
Unset a property on currently deployed application to restoring its default
}
\examples{
\dontrun{

# unset application package cache property to revert to default
unsetProperty("application.package.cache")

}
}

