% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/configureApp.R
\name{unsetProperty}
\alias{unsetProperty}
\title{Unset Application property}
\usage{
unsetProperty(propertyName, appPath = getwd(), appName = NULL,
  account = NULL, force = FALSE)
}
\arguments{
\item{propertyName}{Name of property to unset}

\item{appPath}{Directory or file that was deployed. Defaults to current
working directory.}

\item{appName}{Name of application}

\item{account}{Account name. If a single account is registered on the system
then this parameter can be omitted.}

\item{force}{Forcibly unset the property}
}
\description{
Unset a property on currently deployed ShinyApps application (restoring to
its default value)
}
\note{
This function only works for ShinyApps servers.
}
\examples{
\dontrun{

# unset application package cache property to revert to default
unsetProperty("application.package.cache")

}
}

